Move tests to test_geom_methods.py
For geometry tests, test both the GeoSeries and GeoDataFrame versions.

If the method's input is also a GeoSeries, test the input for both
GeoSeries and GeoDataFrame as well.

For example, for intersects(), test the following combination of inputs:
GeoSeries, GeoSeries,
GeoDataFrame, GeoSeries,
GeoSeries, GeoDataFrame,
GeoDataFrame, GeoDataFrame

Helper functions are provided to test all combinations